At last, Random VS now supports CT-CODE and will exclusively consist of random CTs. Should be ready for testing on all regions! Pretty sure PAL works just as intended, both on Dolphin (through a CTGP v1.02 ISO on PAL) and on my Wii U.
As a side effect, "In Order" pooling will consist of either Nintendo courses or CTs depending on which category the first picked course belongs to. If it's a Nintendo
course, then slot 8.4 wraps to 1.1; otherwise, the last CT wraps back to the first CT.

Also, if the distribution has less than 8 CT cups, the CT pool is adjusted as such.
This makes it impossible for a Random VS to replay the same CT more than once.

Signed-off-by: DimitriPilot3 <DimitriPilot3@users.noreply.github.com>